refactor: move `getVersion` to JsPackageManager
