+++ lib/cli/src/generators/REACT_NATIVE/index.ts
@@ -74,7 +74,7 @@ export default async (
 
   writePackageJson(packageJson);
 
-  const babelDependencies = await getBabelDependencies(npmOptions, packageJson);
+  const babelDependencies = await getBabelDependencies(packageManager, npmOptions, packageJson);
 
   packageManager.addDependencies({ ...npmOptions, packageJson }, [
     ...devDependencies,
