+++ lib/cli/src/generators/RIOT/index.ts
@@ -42,7 +42,7 @@ const generator: Generator = async (packageManager, npmOptions, { storyFormat })
 
   writePackageJson(packageJson);
 
-  const babelDependencies = await getBabelDependencies(npmOptions, packageJson);
+  const babelDependencies = await getBabelDependencies(packageManager, npmOptions, packageJson);
 
   packageManager.addDependencies({ ...npmOptions, packageJson }, [
     ...dependencies,
