+++ lib/cli/src/helpers.ts
@@ -3,46 +3,16 @@ import path from 'path';
 import fs from 'fs';
 import fse from 'fs-extra';
 import chalk from 'chalk';
-import { gt, satisfies } from '@storybook/semver';
+import { satisfies } from '@storybook/semver';
 import stripJsonComments from 'strip-json-comments';
 
 import { latestVersion } from './latest_version';
 import { StoryFormat } from './project_types';
 import { PackageJson } from './PackageJson';
 import { NpmOptions } from './NpmOptions';
-
-// Cannot be `import` as it's not under TS root dir
-const { version, devDependencies } = require('../package.json');
+import { JsPackageManager } from './js-package-manager';
 
 const logger = console;
-
-export async function getVersion(npmOptions: NpmOptions, packageName: string, constraint?: any) {
-  let current;
-  if (packageName === '@storybook/cli') {
-    current = version;
-  } else if (/storybook/.test(packageName)) {
-    current = devDependencies[packageName];
-  }
-
-  let latest;
-  try {
-    latest = await latestVersion(npmOptions, packageName, constraint);
-  } catch (e) {
-    if (current) {
-      logger.warn(`\n     ${chalk.yellow(e.message)}`);
-      return current;
-    }
-
-    logger.error(`\n     ${chalk.red(e.message)}`);
-    process.exit(1);
-  }
-
-  const versionToUse =
-    current && (!constraint || satisfies(current, constraint)) && gt(current, latest)
-      ? current
-      : latest;
-  return `^${versionToUse}`;
-}
 
 export function getPackageJson() {
   const packageJsonPath = path.resolve('package.json');
@@ -160,14 +130,18 @@ export function codeLog(codeLines: string[], leftPadAmount?: number) {
  * @param {Object} packageJson The current package.json so we can inspect its contents
  * @returns {Array} Contains the packages and versions that need to be installed
  * @example
- * const babelDependencies = await getBabelDependencies(npmOptions, packageJson);
+ * const babelDependencies = await getBabelDependencies(packageManager, npmOptions, packageJson);
  * // you can then spread the result when using installDependencies
  * installDependencies(npmOptions, [
  *   `@storybook/react@${storybookVersion}`,
  *   ...babelDependencies,
  * ]);
  */
-export async function getBabelDependencies(npmOptions: NpmOptions, packageJson: PackageJson) {
+export async function getBabelDependencies(
+  packageManager: JsPackageManager,
+  npmOptions: NpmOptions,
+  packageJson: PackageJson
+) {
   const dependenciesToAdd = [];
   let babelLoaderVersion = '^8.0.0-0';
 
@@ -176,7 +150,7 @@ export async function getBabelDependencies(npmOptions: NpmOptions, packageJson: 
 
   if (!babelCoreVersion) {
     if (!packageJson.dependencies['@babel/core'] && !packageJson.devDependencies['@babel/core']) {
-      const babelCoreInstallVersion = await getVersion(npmOptions, '@babel/core');
+      const babelCoreInstallVersion = await packageManager.getVersion(npmOptions, '@babel/core');
       dependenciesToAdd.push(`@babel/core@${babelCoreInstallVersion}`);
     }
   } else {
@@ -192,7 +166,7 @@ export async function getBabelDependencies(npmOptions: NpmOptions, packageJson: 
   }
 
   if (!packageJson.dependencies['babel-loader'] && !packageJson.devDependencies['babel-loader']) {
-    const babelLoaderInstallVersion = await getVersion(
+    const babelLoaderInstallVersion = await packageManager.getVersion(
       npmOptions,
       'babel-loader',
       babelLoaderVersion
